+++ python/GafferSceneUI/__init__.py
@@ -205,6 +205,7 @@ from . import MergePointsUI
 from . import MergeCurvesUI
 from . import VisualiserToolUI
 from . import PrimitiveVariableTweaksUI
+from . import ShuffleRenderPassesUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.